requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
watchfiles>=0.21.0
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
